monitored hosts have more verbose names, attributes may be added
Newly-created host names in opsview now include stack and phase if available, in addition to application and instance id.
If removing a host from monitoring fails, it will fall back to trying to remove the old name of just app and id.

Attributes may be added to a monitored host via a new api call.
Icon Name may also be specified.